[Bug 1286]:  Fix TAPI for VLANs: vid is uint16_t.
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@35291 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/tapi/tapi_cfg_base.h
@@ -316,7 +316,7 @@ tapi_cfg_base_if_arp_disable(const char *ta, const char * iface)
  * @return Status code
  */
 te_errno tapi_cfg_base_if_add_vlan(const char *ta, const char *if_name,
-                                   int vid, char **vlan_ifname);
+                                   uint16_t vid, char **vlan_ifname);
 
 /**
  * Delete VLAN interface.
@@ -328,7 +328,7 @@ te_errno tapi_cfg_base_if_add_vlan(const char *ta, const char *if_name,
  * @return Status code
  */
 te_errno tapi_cfg_base_if_del_vlan(const char *ta, const char *if_name, 
-                                   int vid);
+                                   uint16_t vid);
 
 #ifdef __cplusplus
 } /* extern "C" */
